Replication, trying to create a preauthorization would throw an error when using random primary keys sometimes. Backported to: 14.3.35.0;15.1.13.0 Reviewed by: No one.  Blantant bug with only one possible fix.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/opendental14.3@11976 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Main Modules/ContrTreat.cs
@@ -2972,11 +2972,11 @@ namespace OpenDental{
 
 		///<summary>Similar method in Account</summary>
 		private bool CheckClearinghouseDefaults() {
-			if(PrefC.GetInt(PrefName.ClearinghouseDefaultDent)==0) {
+			if(PrefC.GetLong(PrefName.ClearinghouseDefaultDent)==0) {
 				MsgBox.Show(this,"No default dental clearinghouse defined.");
 				return false;
 			}
-			if(PrefC.GetBool(PrefName.ShowFeatureMedicalInsurance) && PrefC.GetInt(PrefName.ClearinghouseDefaultMed)==0) {
+			if(PrefC.GetBool(PrefName.ShowFeatureMedicalInsurance) && PrefC.GetLong(PrefName.ClearinghouseDefaultMed)==0) {
 				MsgBox.Show(this,"No default medical clearinghouse defined.");
 				return false;
 			}
